Transition to C++11 - Adjusted R and Q values - wip - SLAM tested and runs; featureSet left untested on linux (func should remain same as code was left untouched) - Fixed MKL cmake inclusion - MKL loads on linux OS

+++ c++/src/lib/shared_structs/include/SharedMemoryStructs.h
@@ -84,7 +84,7 @@ typedef struct {
 } POSE;
 
 typedef struct {
-    unsigned long idx{}; // TODO I don't like this here (doesn't relate to feature)
+    uint16_t idx{}; // TODO I don't like this here (doesn't relate to feature)
     float correspondence{};
     RAY incidentRay{};
     POSE pose{};
